move the single command logic into a subcommand, to allow for future additional subcommands
