+++ jte-cli/src/test/java/gg/jte/cli/CliArgsTest.java
@@ -2,6 +2,7 @@ package gg.jte.cli;
 
 import gg.jte.ContentType;
 import org.junit.jupiter.api.Test;
+import picocli.CommandLine;
 
 import java.nio.file.Path;
 
@@ -11,8 +12,8 @@ class CliArgsTest {
 
     @Test
     void parsesAllOptions() {
-        Main main = new Main();
-        Main.newCommandLine(main).parseArgs(
+        GenerateCommand generateCommand = parseGenerate(
+                "generate",
                 "-s", "src/jte",
                 "-t", "target/generated",
                 "-c", "Html",
@@ -24,32 +25,35 @@ class CliArgsTest {
                 "--target-resource-directory", "target/resources"
         );
 
-        assertThat(main.sourceDirectory).isEqualTo(Path.of("src/jte"));
-        assertThat(main.targetDirectory).isEqualTo(Path.of("target/generated"));
-        assertThat(main.contentType).isEqualTo(ContentType.Html);
-        assertThat(main.trimControlStructures).isTrue();
-        assertThat(main.htmlTags).containsExactly("form", "input");
-        assertThat(main.htmlCommentsPreserved).isTrue();
-        assertThat(main.binaryStaticContent).isTrue();
-        assertThat(main.packageName).isEqualTo("com.example.templates");
-        assertThat(main.targetResourceDirectory).isEqualTo(Path.of("target/resources"));
+        assertThat(generateCommand.sourceDirectory).isEqualTo(Path.of("src/jte"));
+        assertThat(generateCommand.targetDirectory).isEqualTo(Path.of("target/generated"));
+        assertThat(generateCommand.contentType).isEqualTo(ContentType.Html);
+        assertThat(generateCommand.trimControlStructures).isTrue();
+        assertThat(generateCommand.htmlTags).containsExactly("form", "input");
+        assertThat(generateCommand.htmlCommentsPreserved).isTrue();
+        assertThat(generateCommand.binaryStaticContent).isTrue();
+        assertThat(generateCommand.packageName).isEqualTo("com.example.templates");
+        assertThat(generateCommand.targetResourceDirectory).isEqualTo(Path.of("target/resources"));
     }
 
     @Test
     void defaultsPackageNameAndRequiresSourceTargetAndContentType() {
-        Main main = new Main();
-        Main.newCommandLine(main).parseArgs("-s", "src/jte", "-t", "target/generated", "-c", "Plain");
+        GenerateCommand generateCommand = parseGenerate("generate", "-s", "src/jte", "-t", "target/generated", "-c", "Plain");
 
-        assertThat(main.packageName).isEqualTo("gg.jte.generated.precompiled");
-        assertThat(main.trimControlStructures).isFalse();
-        assertThat(main.htmlTags).isNull();
+        assertThat(generateCommand.packageName).isEqualTo("gg.jte.generated.precompiled");
+        assertThat(generateCommand.trimControlStructures).isFalse();
+        assertThat(generateCommand.htmlTags).isNull();
     }
 
     @Test
     void parsesContentTypeCaseInsensitively() {
-        Main main = new Main();
-        Main.newCommandLine(main).parseArgs("-s", "src/jte", "-t", "target/generated", "-c", "html");
+        GenerateCommand generateCommand = parseGenerate("generate", "-s", "src/jte", "-t", "target/generated", "-c", "html");
 
-        assertThat(main.contentType).isEqualTo(ContentType.Html);
+        assertThat(generateCommand.contentType).isEqualTo(ContentType.Html);
+    }
+
+    private static GenerateCommand parseGenerate(String... args) {
+        CommandLine.ParseResult result = Main.newCommandLine(new Main()).parseArgs(args);
+        return (GenerateCommand) result.subcommand().commandSpec().userObject();
     }
 }
